+++ application/app/role-teacher/content/cards/CardContentSlide.jsx
@@ -8,7 +8,8 @@ export default function CardContentSlide({ content, onDelete }) {
 
   const getSlideUrl = (filePath) => {
     if (!filePath) return null
-    return `http://localhost:8003/${filePath}`
+    const { CONTENT_API_BASE_URL } = require('@/lib/content/content-api-config');
+    return `${CONTENT_API_BASE_URL}/${filePath}`
   }
 
   const slideUrl = getSlideUrl(content.file_path)
