+++ application/app/role-teacher/content/cards/CardContentVideo.jsx
@@ -11,7 +11,8 @@ export default function CardContentVideo({ content, onDelete }) {
 
   const getVideoUrl = (filePath) => {
     if (!filePath) return null
-    return `http://localhost:8003/${filePath}`
+    const { CONTENT_API_BASE_URL } = require('@/lib/content/content-api-config');
+    return `${CONTENT_API_BASE_URL}/${filePath}`
   }
 
   const videoUrl = getVideoUrl(content.file_path)
