+++ application/env.example
@@ -1,0 +1,5 @@
+NEXT_PUBLIC_API_GATEWAY_URL=http://localhost:8010
+NEXT_PUBLIC_SANDBOX_SERVICE_URL=http://localhost:8002
+NEXT_PUBLIC_CONTENT_SERVICE_URL=http://localhost:8003
+NEXT_PUBLIC_AI_SERVICE_URL=http://localhost:8005
+NEXT_PUBLIC_USER_SERVICE_URL=http://localhost:8006
