+++ application/lib/auth/api.js
@@ -1,4 +1,5 @@
-const BASE_URL = process.env.NEXT_PUBLIC_API_URL || 'http://127.0.0.1:8006';
+const { API_BASE_URL } = require('../users/api-config');
+const BASE_URL = API_BASE_URL;
 
 const apiFetch = {
   post: async (endpoint, data) => {
