+++ application/lib/content/content-api-config.js
@@ -1,4 +1,4 @@
-const CONTENT_API_BASE_URL = 'http://localhost:8010';
+const CONTENT_API_BASE_URL = process.env.NEXT_PUBLIC_API_GATEWAY_URL || 'http://localhost:8010';
 
 const defaultContentHeaders = {
   'Content-Type': 'application/json',
