+++ application/lib/tasks-teacher/task-api-config.js
@@ -1,4 +1,4 @@
-const TASK_API_BASE_URL = 'http://localhost:8010';
+const TASK_API_BASE_URL = process.env.NEXT_PUBLIC_API_GATEWAY_URL || 'http://localhost:8010';
 
 const defaultTaskHeaders = {
   'Content-Type': 'application/json',
