+++ application/lib/users/api-config.js
@@ -1,4 +1,4 @@
-const API_BASE_URL = 'http://localhost:8010';
+const API_BASE_URL = process.env.NEXT_PUBLIC_API_GATEWAY_URL || 'http://localhost:8010';
 
 const defaultHeaders = {
   'Content-Type': 'application/json',
